% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghred_tisefka_UI.R
\name{ghred_tisefka_mod}
\alias{ghred_tisefka_mod}
\title{Saldae dashboard module: upload data}
\usage{
ghred_tisefka_mod(input, output, session)
}
\arguments{
\item{input}{input shinydashboard elements containing information to use data upload}

\item{output}{output shinydashboard element}

\item{session}{shiny session}
}
\value{
output objects to be displayed in corresponding UI module
}
\description{
upload rwa data and prepare it to be used for exploration and analysis
}
\author{
Farid Azouaou
}
