% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghred_tisefka_UI.R
\name{ghred_tisefka_UI}
\alias{ghred_tisefka_UI}
\title{Saldae Dashboard Module UI}
\usage{
ghred_tisefka_UI(id)
}
\arguments{
\item{id}{server module ID}
}
\description{
Saldae Dashboard module UI : data upload
}
\author{
Farid Azouaou
}
