% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tisefka_UI.R
\name{mod_sekned_tisefka_iceqfan}
\alias{mod_sekned_tisefka_iceqfan}
\title{Saldae Dashboard Module plotly rawdata}
\usage{
mod_sekned_tisefka_iceqfan(
  tisefka = NULL,
  target_variables = NULL,
  graph_type = NULL
)
}
\arguments{
\item{tisefka}{raw data}

\item{target_variables}{variables to plot}

\item{graph_type}{plot type  "bar" "scatter" "line" "hist",..}
}
\value{
plotly interactive object
}
\description{
multiple plots : data x-axis date and y-axis numerical variable
}
\author{
Farid Azouaou
}
