% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tisefka_UI.R
\name{mod_sekned_yiwet_tisefka}
\alias{mod_sekned_yiwet_tisefka}
\title{Saldae Dashboard Module plotly rawdata single}
\usage{
mod_sekned_yiwet_tisefka(
  tisefka = NULL,
  variable_inu = NULL,
  graph_type = "scatter",
  graph_col = "#228B22"
)
}
\arguments{
\item{tisefka}{raw data}

\item{variable_inu}{variable to plot}

\item{graph_type}{plot type  "bar" "scatter" "line" "hist",..}
}
\value{
plotly interactive object
}
\description{
Plot single variable
}
\author{
Farid Azouaou
}
