% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tisefka_UI.R
\name{SA_aggregation_funs}
\alias{SA_aggregation_funs}
\title{Saldae Aggregation functions}
\usage{
SA_aggregation_funs(aggregation_metric = NULL)
}
\arguments{
\item{aggregation_metric}{aggregation metric :Average , Sum, Min , Max,}
}
\description{
Plot single variable
}
\author{
Farid Azouaou
}
