% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tisefka_anomaly_UI.R
\name{SA_anomaly_UI}
\alias{SA_anomaly_UI}
\title{Saldae Dashboard Module Anomaly Detection Advanced}
\usage{
SA_anomaly_UI(
  id,
  mod_title = NULL,
  div_width = "col-xs-12 col-sm-12 col-md-12"
)
}
\arguments{
\item{id}{t.b.d}
}
\description{
Saldae Dashboard Module Anomaly Detection Advanced
}
\author{
Farid Azouaou
}
