% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_Value_Box_UI.R
\name{SA_key_figures_Server}
\alias{SA_key_figures_Server}
\title{Saldae Dashboard Module Server}
\usage{
SA_key_figures_Server(input, output, session, tisefka)
}
\arguments{
\item{input}{input shinydashboard elements containing information to use for output generation}

\item{output}{output shinydashboard element}

\item{session}{shiny session}

\item{tisefka}{reactive object containing  value data}
}
\value{
output objects to be displayed in corresponding UI module
}
\description{
Saldae Dashboard module SERVER : render and generate 3 value boxes output
}
