% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_UI.R
\name{SA_reporting_mod}
\alias{SA_reporting_mod}
\title{reporting_board_module Server Function}
\usage{
SA_reporting_mod(input, output, session, tisefka_list)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\author{
Farid Azouaou
}
