% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_UI.R
\name{SA_reporting_UI}
\alias{SA_reporting_UI}
\title{reporting_board_module UI Function}
\usage{
SA_reporting_UI(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\author{
Farid Azouaou
}
