% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate_UI.R
\name{SA_tisefka_gemmu_UI}
\alias{SA_tisefka_gemmu_UI}
\title{Saldae Dashboard Module UI (growth rate)}
\usage{
SA_tisefka_gemmu_UI(
  id,
  mod_title = NULL,
  div_width = "col-xs-12 col-sm-6 col-md-8"
)
}
\arguments{
\item{id}{server module ID}

\item{mod_title}{module title (default NULL)}

\item{div_width}{dimension information about the framework(html object)}
}
\value{
UI module
}
\description{
Saldae Dashboard module UI : growth rate calculator
}
\author{
Farid Azouaou
}
