% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tisefka_UI.R
\name{SA_tisefka_mod}
\alias{SA_tisefka_mod}
\title{Saldae Dashboard Module Server}
\usage{
SA_tisefka_mod(input, output, session, tisefka)
}
\arguments{
\item{input}{input shinydashboard elements containing information to use for output generation}

\item{output}{output shinydashboard element}

\item{session}{shiny session}

\item{tisefka}{reactive object containing data}
}
\value{
output objects to be displayed in corresponding UI module
}
\description{
Saldae Dashboard module SERVER : render and generate object to be displayed data(chart/table)
}
\author{
Farid Azouaou
}
