% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tisefka_overview_UI.R
\name{SA_tisefka_overview_server}
\alias{SA_tisefka_overview_server}
\title{Saldae Dashboard Data Overview Server module}
\usage{
SA_tisefka_overview_server(input, output, session, tisefka)
}
\arguments{
\item{input}{input shinydashboard elements containing information to use for output generation}

\item{output}{output shinydashboard element}

\item{session}{shiny session}

\item{tisefka}{reactive object containing  value data}
}
\value{
output objects to be displayed in corresponding UI module
}
\description{
Saldae Dashboard module SERVER : render and generate 3 value boxes output
}
