% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tisefka_overview_UI.R
\name{SA_tisefka_overview_UI}
\alias{SA_tisefka_overview_UI}
\title{Saldae Dashboard Data Overview Server UI}
\usage{
SA_tisefka_overview_UI(id)
}
\arguments{
\item{id}{server module ID}
}
\value{
n  value boxes objects
}
\description{
Saldae Dashboard module UI :  value boxes  in a row containing
}
\author{
Farid Azouaou
}
