% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tisefka_UI.R
\name{SA_tisefka_UI}
\alias{SA_tisefka_UI}
\title{Saldae Dashboard Module UI}
\usage{
SA_tisefka_UI(
  id,
  mod_title = NULL,
  div_width = "col-xs-12 col-sm-12 col-md-12"
)
}
\arguments{
\item{id}{server module ID}

\item{mod_title}{module title (default NULL)}

\item{div_width}{dimension information about the framework(html object)}
}
\value{
UI module
}
\description{
Saldae Dashboard module UI : display data(chart/table)
}
\author{
Farid Azouaou
}
