% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Value_Box_UI.R
\name{SA_Value_box_UI}
\alias{SA_Value_box_UI}
\title{Saldae Dashboard Module UI: value boxes}
\usage{
SA_Value_box_UI(id)
}
\arguments{
\item{id}{server module ID}
}
\value{
n  value boxes objects
}
\description{
Saldae Dashboard module UI :  value boxes  in a row containing
}
\author{
Farid Azouaou
}
