% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_UI.R
\name{Saldae_amadal_UI}
\alias{Saldae_amadal_UI}
\title{Saldae Dashboard Module UI (Map)}
\usage{
Saldae_amadal_UI(id = NULL, mod_title = NULL)
}
\arguments{
\item{id}{server module ID}

\item{mod_title}{module title (default NULL)}
}
\value{
UI module
}
\description{
Saldae Dashboard module UI : Map  View
}
\author{
Farid Azouaou
}
