% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_UI.R
\name{Saldae_kefrida_UI}
\alias{Saldae_kefrida_UI}
\title{Saldae Dashboard Module UI (Waterfall)}
\usage{
Saldae_kefrida_UI(id = NULL, mod_title = NULL)
}
\arguments{
\item{id}{server module ID}

\item{mod_title}{module title (default NULL)}
}
\value{
UI module
}
\description{
Saldae Dashboard module UI : Waterfall View
}
\author{
Farid Azouaou
}
