% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_UI.R
\name{Saldae_leaflet_amedya}
\alias{Saldae_leaflet_amedya}
\title{Saldae Map based  Graph}
\usage{
Saldae_leaflet_amedya(tisefka = NULL)
}
\arguments{
\item{tisefka}{data containing target region infrmation}
}
\value{
leaflet map object
}
\description{
Saldae Map based  Graph
}
\author{
Farid Azouaou
}
