% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlestick_UI.R
\name{Saldae_taftilt_UI}
\alias{Saldae_taftilt_UI}
\title{Saldae Dashboard Module UI (CandleStick)}
\usage{
Saldae_taftilt_UI(id = NULL, mod_title = NULL)
}
\arguments{
\item{id}{server module ID}

\item{mod_title}{module title (default NULL)}
}
\value{
UI module
}
\description{
Saldae Dashboard module UI : CandleStick View
}
\author{
Farid Azouaou
}
