% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_UI.R
\name{sekned_kefrida_f}
\alias{sekned_kefrida_f}
\title{Saldae Waterfall Graph}
\usage{
sekned_kefrida_f(tisefka = NULL, kefrida_variables = NULL, aggregation = "sum")
}
\arguments{
\item{tisefka}{data}

\item{kefrida_variables}{variables to use}

\item{aggregation}{aggregation logic}
}
\value{
plotly waterfall object
}
\description{
Saldae Waterfall Graph
}
\author{
Farid Azouaou
}
