% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlestick_UI.R
\name{sekned_taftilt_tisefka_f}
\alias{sekned_taftilt_tisefka_f}
\title{Saldae CandleStick Graph}
\usage{
sekned_taftilt_tisefka_f(tisefka = NULL)
}
\arguments{
\item{tisefka}{data containing low,high,close,open  variables}
}
\value{
plotly candlestick object
}
\description{
Saldae CandleStick Graph
}
\author{
Farid Azouaou
}
